pyproject.toml
[project]
name = "daily-report"
version = "0.1.0"
description = "Agente redatto alla scrittura di mail utilizzando lo storico dati di GA4"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datapizza-ai>=0.0.7",
    "datapizza-ai-clients-anthropic>=0.0.4",
    "google-analytics-data>=0.19.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.1",
    "redis>=5.0.0",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.8.2",
    "datapizza-ai-clients-openai>=0.0.7",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
]
